+++ CHANGELOG.md
@@ -4,6 +4,10 @@ FEATURES:
 
 * Support for changing the default Vault address and Kubernetes mount path via CLI flag to the vault-csi-provider binary
 
+BUGS:
+
+* Added missing error handling when transforming SecretProviderClass config to a Vault request [[GH-97](https://github.com/hashicorp/vault-csi-provider/pull/97)]
+
 ## 0.2.0 (April 14th, 2021)
 
 FEATURES: